#3 Creation of DeliveryCalculationService and ExtraFeeCalculationService to calculate delivery fee by given data of the city and the vehicle type.

+++ src/main/resources/db/migration/V3__create_table_weather_conditions.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS weather_conditions(
     id BIGINT NOT NULL AUTO_INCREMENT,
-    station_name  VARCHAR(100),
-    air_temperature INT,
-    wind_speed INT,
-    weather_phenomenon  VARCHAR(100),
+    station_name VARCHAR(100),
+    air_temperature FLOAT,
+    wind_speed FLOAT,
+    weather_phenomenon VARCHAR(100),
     observation_time DATETIME2,
     PRIMARY KEY (id)
 );
